# feat: add `optimize` options