<!--- Provide a general summary of your changes in the title above -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->

- [ ] Chore (a non-breaking change which is related to package maintenance)
- [ ] Bug fix (a non-breaking change which fixes an issue)
- [x] New feature (a non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Description

This PR adds a new `optimize` option to `createClient()` and all client methods that make requests to the Prismic API.

```typescript
const client = createClient({
  optimize: {
    /**
     * Determines if fetching repository metadata should be optimized. The
     * optimizations result in faster queries, less network requests, and more
     * accurate cache-busting.
     *
     * Only opt out of this optimization if you know what you are doing.
     *
     * @defaultValue `true`
     */
    repositoryRequests: true,

    /**
     * Determines if concurrent requests to the same URL should be optimized.
     * The optimizations result in less network requests and quicker responses.
     *
     * Only opt out of this optimization if you know what you are doing.
     *
     * @defaultValue `true`
     */
    concurentRequests: true,
  },
});

// `optimize` can be overridden on a per-request basis.
const page = await client.getByUID("page", "home", {
  optimize: {
    repositoryRequests: false,
    concurentRequests: false,
  },
});
```

All current optimizations are enabled by default.

- `repositoryRequests`: Caches successful requests to `/api/v2` internally for 5 seconds and forces external network caches to not cache `/api/v2` requests for 5 seconds.
- `concurrentRequests`: De-duplicates idential concurrent requests.

All optimizations except `repositoryRequests`'s "force external network caches to not cache" were already in place and were not configurable. This PR allows for explicitly opting out of these optimizations, which should only be done in very specific cases.

### Forcing external network caches to not cache `/api/v2`

A new optimization is introduced in this PR: requests to `/api/v2` are forcibly cache-busted by default after 5 seconds.

This is done by adding a per-request-unique URL parameter (`x-valid-until`) containing an epoch timestamp produced at the time of the request. The timestamp represents the time at which the request is no longer valid.

```
https://example-prismic-repo.cdn.prismic.io/api/v2?x-valid-until=1682556195000
```

The 5 second lifespan is not guaranteed. It is calculated by taking the timestamp at the time of the request, rounding down to the nearest 5 seconds, then adding 5 seconds. Treat it as an unguaranteed optimization that may not always be respected.

Only `/api/v2` requests are optimized. All `/api/v2/documents/search` requests are inherently cache-busted by the `ref` parameter.

The parameter is ignored within the concurrent requests job queue when `optimize.concurrentRequests` is enabled.

This new optimization was prompted by the need to forcibly opt out of [Next.js's new aggressive App Router `fetch()` caching](https://beta.nextjs.org/docs/data-fetching/caching).

### To discuss

1. **Is this PR over-engineered?**
   We technically only need to add the `x-valid-until` parameter to every network request with the option to opt out. The other `optimize` options are nice-to-haves to cover edge cases.

2. **Is `x-valid-until` the correct parameter name?**

3. **Is `optimize` the correct option name?**
   It was inspired by Vite's `optimizeDeps` option.

## Checklist:

<!--- Put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->

- [x] My change requires an update to the official documentation.
- [x] All [TSDoc](https://tsdoc.org) comments are up-to-date and new ones have been added where necessary.
- [x] All new and existing tests are passing.

<!--- A cute animal (or car) picture is welcome to close your PR! -->

🦥