@angeloashmore 2023-04-26
I think there's a bug somewhere in the query-level `optimize` overrides. The latest commit's tests show the inconsistencies.

I'll do some digging to figure out what's going on.